Upgrade/Install: Update sodium_compat to v1.17.1.
The latest version of sodium_compat includes further improvements for PHP 8.1 compatibility.

Release notes:
https://github.com/paragonie/sodium_compat/releases/tag/v1.17.1

A full list of changes in this update can be found on GitHub:
https://github.com/paragonie/sodium_compat/compare/v1.17.0...v1.17.1

Follow-up to [49741], [51002], [51591].

Props jrf, paragoninitiativeenterprises.
Fixes #55453.
Built from https://develop.svn.wordpress.org/trunk@52988


git-svn-id: http://core.svn.wordpress.org/trunk@52577 1a063a9b-81f0-0310-95a4-ce76da25c4cd

+++ wp-includes/sodium_compat/src/Core32/Int64.php
@@ -337,9 +337,9 @@ class ParagonIE_Sodium_Core32_Int64
             $a_i = $a[$i];
             for ($j = 0; $j < $a_l; ++$j) {
                 $b_j = $b[$j];
-                $product = ($a_i * $b_j) + $r[$i + $j];
-                $carry = ($product >> $baseLog2 & 0xffff);
-                $r[$i + $j] = ($product - (int) ($carry * $base)) & 0xffff;
+                $product = (($a_i * $b_j) + $r[$i + $j]);
+                $carry = (((int) $product >> $baseLog2) & 0xffff);
+                $r[$i + $j] = ((int) $product - (int) ($carry * $base)) & 0xffff;
                 $r[$i + $j + 1] += $carry;
             }
         }
